Move manange & save location types action, JSP
